@@ -9,10 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@material-ui/styles": "^4.11.5",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
     "autoprefixer": "^10.2.5",
     "cheerio": "^1.0.0-rc.5",
     "cookie": "^0.4.1",
@@ -21,22 +21,24 @@
     "date-fns": "^2.30.0",
     "iconv-lite": "^0.6.3",
     "lodash": "^4.17.21",
-    "next": "^12.3.4",
+    "next": "^13.5.6",
     "next-universal-cookie": "^3.1.0",
     "node-fetch": "^2.6.1",
     "postcss": "^8.2.8",
-    "prettier": "^2.2.1",
     "prettier-eslint": "^12.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-ga": "^3.3.1",
-    "react-slick": "^0.28.1",
+    "react-slick": "^0.30.2",
     "set-cookie-parser": "^2.4.8",
     "tailwindcss": "^2.0.3",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
     "@types/material-ui": "^0.21.17",
-    "eslint-config-next": "^12.3.4"
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^13.5.6",
+    "eslint-config-prettier": "^9.1.0",
+    "prettier": "^3.2.5"
   }
 }
