@@ -1,24 +1,24 @@
-import React, { useEffect } from 'react'
-import UserContext from '../utils/providers/UserProvider'
-import Head from 'next/head'
-import { ThemeProvider } from '@material-ui/core/styles'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import usePageTracking from '../utils/usePageTracking'
+import React, { useEffect } from "react";
+import UserContext from "../utils/providers/UserProvider";
+import Head from "next/head";
+import { ThemeProvider } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+import usePageTracking from "../utils/usePageTracking";
 
-import theme from '../utils/theme'
+import theme from "../utils/theme";
 
-import '../styles/globals.css'
+import "../styles/globals.css";
 // import 'tailwindcss/tailwind.css'
 
 function App({ Component, pageProps }) {
-  usePageTracking()
+  usePageTracking();
 
   useEffect(() => {
-    const jssStyles = document.querySelector('#jss-server-side')
+    const jssStyles = document.querySelector("#jss-server-side");
     if (jssStyles) {
-      jssStyles.parentElement.removeChild(jssStyles)
+      jssStyles.parentElement.removeChild(jssStyles);
     }
-  }, [])
+  }, []);
 
   return (
     <UserContext>
@@ -31,7 +31,7 @@ function App({ Component, pageProps }) {
         <Component {...pageProps} />
       </ThemeProvider>
     </UserContext>
-  )
+  );
 }
 
-export default App
+export default App;
