@@ -1,7 +1,7 @@
-import React from 'react'
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-import { ServerStyleSheets } from '@material-ui/core/styles'
-import theme from '../utils/theme'
+import React from "react";
+import Document, { Html, Head, Main, NextScript } from "next/document";
+// import { ServerStyleSheets } from '@mui/material/styles'
+import theme from "../utils/theme";
 
 export default class MyDocument extends Document {
   render() {
@@ -9,31 +9,37 @@ export default class MyDocument extends Document {
       <Html lang="en">
         <Head>
           <meta name="theme-color" content={theme.palette.primary.main} />
-          <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:300,400,500,700&display=swap" />
-          <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto+Slab:300,400,500,700&display=swap" />
+          <link
+            rel="stylesheet"
+            href="https://fonts.googleapis.com/css?family=Lato:300,400,500,700&display=swap"
+          />
+          <link
+            rel="stylesheet"
+            href="https://fonts.googleapis.com/css?family=Roboto+Slab:300,400,500,700&display=swap"
+          />
         </Head>
         <body>
           <Main />
           <NextScript />
         </body>
       </Html>
-    )
+    );
   }
 }
 
-MyDocument.getInitialProps = async (ctx) => {
-  const sheets = new ServerStyleSheets()
-  const originalRenderPage = ctx.renderPage
+// MyDocument.getInitialProps = async (ctx) => {
+//   const sheets = new ServerStyleSheets()
+//   const originalRenderPage = ctx.renderPage
 
-  ctx.renderPage = () =>
-    originalRenderPage({
-      enhanceApp: (App) => (props) => sheets.collect(<App {...props} />)
-    })
+//   ctx.renderPage = () =>
+//     originalRenderPage({
+//       enhanceApp: (App) => (props) => sheets.collect(<App {...props} />)
+//     })
 
-  const initialProps = await Document.getInitialProps(ctx)
+//   const initialProps = await Document.getInitialProps(ctx)
 
-  return {
-    ...initialProps,
-    styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()]
-  }
-}
+//   return {
+//     ...initialProps,
+//     styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()]
+//   }
+// }
