@@ -1,35 +1,18 @@
 /* eslint-disable react/no-unknown-property */
 
-import React, { useRef } from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
-import Image from 'next/image'
-import Slider from 'react-slick'
+import React, { useRef } from "react";
+import Head from "next/head";
+import Link from "next/link";
+import Slider from "react-slick";
 
-import { Box, Typography, Grid, Button, Divider } from '@material-ui/core'
-import { withStyles } from '@material-ui/core/styles'
+import { Box, Typography, Grid, Button, Divider } from "@mui/material";
 
-import { getClassifiedDetail } from '../../../utils/getters'
-import Header from '../../../components/Header'
-import Breadcrumb from '../../../components/Breadcrumb'
+import { getClassifiedDetail } from "../../../utils/getters";
+import Header from "../../../components/Header";
+import Breadcrumb from "../../../components/Breadcrumb";
 
-export default withStyles({
-  slider: {
-    '& .slick-dots': {
-      position: 'static'
-    }
-  },
-  slideText: {
-    margin: '8px 0'
-  },
-  slideBox: {
-    outline: 'none',
-    '&:focus': {
-      outline: 'none'
-    }
-  }
-})(({ classes, data }) => {
-  console.log(data)
+export default function Main({ data }) {
+  console.log(data);
 
   const sliderSettings = {
     dots: true,
@@ -38,21 +21,30 @@ export default withStyles({
     arrows: false,
     slidesToShow: 1,
     slidesToScroll: 1,
-    adaptiveHeight: true
-  }
+    adaptiveHeight: true,
+  };
 
-  const sliderRef = useRef()
+  const sliderRef = useRef();
 
   const gotoNext = () => {
-    sliderRef.current.slickNext()
-  }
+    sliderRef.current.slickNext();
+  };
 
   return (
     <Box p={{ xs: 0, md: 1 }}>
       <Head>
         <title>{data.base.title}</title>
-        <link rel="stylesheet" type="text/css" charset="UTF-8" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css" />
-        <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css" />
+        <link
+          rel="stylesheet"
+          type="text/css"
+          charset="UTF-8"
+          href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css"
+        />
+        <link
+          rel="stylesheet"
+          type="text/css"
+          href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"
+        />
       </Head>
 
       <Header data={data} selected="Home" />
@@ -66,32 +58,53 @@ export default withStyles({
           <Breadcrumb crumbs={data.page.nav} />
         </Box>
         <Box mb={2}>
-          {data.page.mainPhoto.src
-            ? (
-              <Slider {...sliderSettings} ref={sliderRef} className={classes.slider}>
-                {!data.page.thumbnails.length && (
-                  <Box onClick={gotoNext} className={classes.slideBox}>
-                    <img src={data.page.mainPhoto.src} alt={data.page.mainPhoto.alt} width="100%" />
-                  </Box>
-                )}
-                {data.page.thumbnails.map((item, i) => (
-                  <Box key={`detail-image-${data.page.adId}-${i}`} onClick={gotoNext} className={classes.slideBox}>
-                    <img src={item.src.replace('thumbnails/', '')} alt={item.alt} width="100%" />
-                    {item.label && (
-                      <Typography variant="body2" align="center" className={classes.slideText}>
-                        {item.label}
-                      </Typography>
-                    )}
-                  </Box>
-                ))}
-              </Slider>
-            )
-            : (
-              <Divider />
-            )}
+          {data.page.mainPhoto.src ? (
+            <Slider
+              {...sliderSettings}
+              ref={sliderRef}
+              sx={{ "& .slick-dots": { position: "static" } }}
+            >
+              {!data.page.thumbnails.length && (
+                <Box onClick={gotoNext} className={classes.slideBox}>
+                  <img
+                    src={data.page.mainPhoto.src}
+                    alt={data.page.mainPhoto.alt}
+                    width="100%"
+                  />
+                </Box>
+              )}
+              {data.page.thumbnails.map((item, i) => (
+                <Box
+                  key={`detail-image-${data.page.adId}-${i}`}
+                  onClick={gotoNext}
+                  sx={{ outline: "none", "&:focus": { outline: "none" } }}
+                >
+                  <img
+                    src={item.src.replace("thumbnails/", "")}
+                    alt={item.alt}
+                    width="100%"
+                  />
+                  {item.label && (
+                    <Typography
+                      variant="body2"
+                      align="center"
+                      sx={{ margin: "8px 0" }}
+                    >
+                      {item.label}
+                    </Typography>
+                  )}
+                </Box>
+              ))}
+            </Slider>
+          ) : (
+            <Divider />
+          )}
         </Box>
         <Box mb={2}>
-          <Typography component="div" dangerouslySetInnerHTML={{ __html: data.page.description }} />
+          <Typography
+            component="div"
+            dangerouslySetInnerHTML={{ __html: data.page.description }}
+          />
         </Box>
         <Box mb={1} border={1} borderColor="secondary.light">
           <Box bgcolor="secondary.light" p={1}>
@@ -102,22 +115,34 @@ export default withStyles({
               <Grid item xs={4}>
                 <Box pr={1}>
                   {data.page.advertiserInfo.labels.map((label, i) => (
-                    <Typography key={`advertiserInfo-label-${data.page.adId}-${i}`}>{label}</Typography>
+                    <Typography
+                      key={`advertiserInfo-label-${data.page.adId}-${i}`}
+                    >
+                      {label}
+                    </Typography>
                   ))}
                 </Box>
               </Grid>
               <Grid item xs={8}>
                 <Link href={data.page.advertiserInfo.memberHref} passHref>
-                  <Typography component="a">{data.page.advertiserInfo.member}</Typography>
+                  {data.page.advertiserInfo.member}
                 </Link>
                 {data.page.advertiserInfo.textValues.map((value, i) => (
-                  <Typography key={`advertiserInfo-textValue-${data.page.adId}-${i}`}>{value}</Typography>
+                  <Typography
+                    key={`advertiserInfo-textValue-${data.page.adId}-${i}`}
+                  >
+                    {value}
+                  </Typography>
                 ))}
               </Grid>
               {data.page.advertiserInfo.contactPhone && (
                 <Grid item>
                   <Box mt={2} pr={2}>
-                    <Button variant="outlined" component="a" href={`tel:${data.page.advertiserInfo.contactPhone}`}>
+                    <Button
+                      variant="outlined"
+                      component="a"
+                      href={`tel:${data.page.advertiserInfo.contactPhone}`}
+                    >
                       Call
                     </Button>
                   </Box>
@@ -146,12 +171,18 @@ export default withStyles({
               <Grid container key={`detail-ad-info-${data.page.adId}-${i}`}>
                 <Grid item xs={12} sm={3} md={2}>
                   <Box pr={1}>
-                    <Typography component="div" dangerouslySetInnerHTML={{ __html: el[0] }} />
+                    <Typography
+                      component="div"
+                      dangerouslySetInnerHTML={{ __html: el[0] }}
+                    />
                   </Box>
                 </Grid>
                 <Grid item xs={12} sm={9} md={10}>
                   <Box pl={2} pb={[1, 0]}>
-                    <Typography component="div" dangerouslySetInnerHTML={{ __html: el[1] }} />
+                    <Typography
+                      component="div"
+                      dangerouslySetInnerHTML={{ __html: el[1] }}
+                    />
                   </Box>
                 </Grid>
               </Grid>
@@ -160,16 +191,16 @@ export default withStyles({
         </Box>
       </Box>
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getClassifiedDetail(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getClassifiedDetail(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
