@@ -1,19 +1,21 @@
-import Head from 'next/head'
-import React from 'react'
-import Link from 'next/link'
+import Head from "next/head";
+import React from "react";
+import Link from "next/link";
 
-import { Box, Grid, Typography, Hidden, Divider } from '@material-ui/core'
-import { withStyles } from '@material-ui/core/styles'
+import { Box, Grid, Typography, Hidden, Divider } from "@mui/material";
+import { styled } from "@mui/system";
+import { useTheme } from "@mui/material/styles";
 
-import { getForums } from '../../../utils/getters'
-import Header from '../../../components/Header'
-import theme from '../../../utils/theme'
+import { getForums } from "../../../utils/getters";
+import Header from "../../../components/Header";
 
-export default withStyles({
-  root: {
-    color: 'white'
-  }
-})(({ data, classes }) => {
+const StyledTypography = styled(Typography)({
+  color: "white",
+});
+
+export default function Main({ data }) {
+  const theme = useTheme();
+
   return (
     <Box p={{ xs: 0, md: 1 }}>
       <Head>
@@ -28,30 +30,22 @@ export default withStyles({
             <Grid container>
               <Grid item xs={8}>
                 <Box p={1}>
-                  <Typography align="center" className={classes.root}>
-                    Forum
-                  </Typography>
+                  <StyledTypography align="center">Forum</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={1}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Topics
-                  </Typography>
+                  <StyledTypography align="right">Topics</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={1}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Posts
-                  </Typography>
+                  <StyledTypography align="right">Posts</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={2}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Last Post
-                  </Typography>
+                  <StyledTypography align="right">Last Post</StyledTypography>
                 </Box>
               </Grid>
             </Grid>
@@ -61,23 +55,30 @@ export default withStyles({
         {data.page.forumGroups.map((el, i) => (
           <Box mb={2} key={`forumGroup-${i}`}>
             <Box p={1} bgcolor="secondary.light">
-              <Link href={el.href} passHref>
-                <Typography component="a" variant="subtitle1">
-                  {el.title}
-                </Typography>
+              <Link href={el.href} passHref variant="subtitle1">
+                {el.title}
               </Link>
             </Box>
             {el.items.map((subEl, subI) => (
-              <Box py={1} px={1} mb={{ xs: 0, sm: 1 }} borderBottom={1} style={{ borderColor: theme.palette.secondary.light }} key={`forum-${i}-${subI}`}>
+              <Box
+                py={1}
+                px={1}
+                mb={{ xs: 0, sm: 1 }}
+                borderBottom={1}
+                style={{ borderColor: theme.palette.secondary.light }}
+                key={`forum-${i}-${subI}`}
+              >
                 <Grid container>
                   <Grid item xs={12} sm={8}>
                     <Box mb={{ xs: 0, sm: 0 }}>
                       {/* {subEl.newPosts  ? '!' : '-'} */}
 
-                      <Link href={subEl.href} passHref>
-                        <Typography component="a" className="font-semibold">
-                          {subEl.title}
-                        </Typography>
+                      <Link
+                        href={subEl.href}
+                        passHref
+                        className="font-semibold"
+                      >
+                        {subEl.title}
                       </Link>
                       <Hidden xsDown>
                         <Typography>{subEl.description}</Typography>
@@ -86,13 +87,19 @@ export default withStyles({
                   </Grid>
                   <Hidden smUp>
                     <Grid item xs={12} container justifyContent="flex-end">
-                      <Typography align="right">Topics: {subEl.topics}</Typography>
+                      <Typography align="right">
+                        Topics: {subEl.topics}
+                      </Typography>
                       <Box mx={1} height={16}>
                         <Divider orientation="vertical" />
                       </Box>
-                      <Typography align="right">Posts: {subEl.posts}</Typography>
+                      <Typography align="right">
+                        Posts: {subEl.posts}
+                      </Typography>
                       <Grid item xs={12}>
-                        <Typography align="right">Last Post: {subEl.lastPost.text}</Typography>
+                        <Typography align="right">
+                          Last Post: {subEl.lastPost.text}
+                        </Typography>
                       </Grid>
                     </Grid>
                   </Hidden>
@@ -109,7 +116,9 @@ export default withStyles({
                     </Grid>
                     <Grid item xs={6} sm={2}>
                       <Box pl={1}>
-                        <Typography align="right">{subEl.lastPost.text}</Typography>
+                        <Typography align="right">
+                          {subEl.lastPost.text}
+                        </Typography>
                       </Box>
                     </Grid>
                   </Hidden>
@@ -120,16 +129,16 @@ export default withStyles({
         ))}
       </Box>
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getForums(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getForums(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
