@@ -1,36 +1,20 @@
-import React from 'react'
-import Head from 'next/head'
-// import Link from 'next/link'
-import Image from 'next/image'
-import copy from 'copy-to-clipboard'
+import React from "react";
+import Head from "next/head";
+import copy from "copy-to-clipboard";
 
-import { Box, Typography, Grid } from '@material-ui/core'
-import { withStyles } from '@material-ui/core/styles'
+import { Box, Typography, Grid } from "@mui/material";
 
-import { getGalleryPage } from '../../../utils/getters'
-import Header from '../../../components/Header'
-import Breadcrumb from '../../../components/Breadcrumb'
+import { getGalleryPage } from "../../../utils/getters";
+import Header from "../../../components/Header";
+import Breadcrumb from "../../../components/Breadcrumb";
 
-export default withStyles({
-  root: {
-    color: 'white'
-  },
-  topLinkSpacing: {
-    marginTop: '8px',
-    padding: '8px 0'
-  },
-  image: {
-    width: 'auto',
-    maxWidth: '100%',
-    height: 'auto'
-  }
-})(({ data, classes }) => {
+export default function Main({ data }) {
   const handleCopy = (e) => {
     copy(e.target.dataset.copy, {
       debug: true,
-      message: 'Press #{key} to copy'
-    })
-  }
+      message: "Press #{key} to copy",
+    });
+  };
 
   return (
     <Box p={{ xs: 0, md: 1 }}>
@@ -47,7 +31,7 @@ export default withStyles({
 
         <Box bgcolor="primary.main">
           <Box p={1} px={2}>
-            <Typography align="center" className={classes.root}>
+            <Typography align="center" sx={{ color: "white" }}>
               {data.page.title}
             </Typography>
           </Box>
@@ -56,12 +40,22 @@ export default withStyles({
         <Box mb={2}>
           <Grid container>
             <Grid item xs={6} sm={4} md={2}>
-              <Typography align="center" data-copy={data.page.photoLink} onClick={handleCopy} className={classes.topLinkSpacing}>
+              <Typography
+                align="center"
+                data-copy={data.page.photoLink}
+                onClick={handleCopy}
+                sx={{ marginTop: "8px", padding: "8px 0" }}
+              >
                 Copy photo link
               </Typography>
             </Grid>
             <Grid item xs={6}>
-              <Typography align="center" data-copy={data.page.forumCode} onClick={handleCopy} className={classes.topLinkSpacing}>
+              <Typography
+                align="center"
+                data-copy={data.page.forumCode}
+                onClick={handleCopy}
+                sx={{ marginTop: "8px", padding: "8px 0" }}
+              >
                 Copy forum link
               </Typography>
             </Grid>
@@ -69,16 +63,25 @@ export default withStyles({
               <Box p={1}>
                 {/* <Link href={data.page.href} passHref>
                   <a style={{ display: 'block' }}> */}
-                <img src={data.page.img.src} alt={data.page.img.alt} className={classes.image} />
+                <img
+                  src={data.page.img.src}
+                  alt={data.page.img.alt}
+                  sx={{ width: "auto", maxWidth: "100%", height: "auto" }}
+                />
                 {/* <div style={{ position: 'relative', width: '100%', height: '400px', paddingBottom: '20%' }}>
-                  <Image src={data.page.img.src} alt={data.page.img.alt} layout="fill" objectFit="cover" />
+                  <Image src={data.page.img.src} alt={data.page.img.alt} fill={true} style={{ objectFit: "cover" }} />
                 </div> */}
                 {/* </a>
                 </Link> */}
               </Box>
               <Box p={1}>
                 <Typography>{data.page.photo.title}</Typography>
-                <Typography component="div" dangerouslySetInnerHTML={{ __html: data.page.photo.description }} />
+                <Typography
+                  component="div"
+                  dangerouslySetInnerHTML={{
+                    __html: data.page.photo.description,
+                  }}
+                />
               </Box>
             </Grid>
             <Grid item xs={12}>
@@ -90,7 +93,10 @@ export default withStyles({
                   <Typography>
                     <strong>{data.page.author.name}</strong>
                   </Typography>
-                  <Typography component="div" dangerouslySetInnerHTML={{ __html: data.page.author.info }} />
+                  <Typography
+                    component="div"
+                    dangerouslySetInnerHTML={{ __html: data.page.author.info }}
+                  />
                 </Box>
               </Box>
             </Grid>
@@ -98,16 +104,16 @@ export default withStyles({
         </Box>
       </Box>
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getGalleryPage(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getGalleryPage(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
