@@ -1,19 +1,12 @@
-import React from 'react'
-import Head from 'next/head'
-import Image from 'next/image'
+import React from "react";
+import Head from "next/head";
 
-import { Box, Typography, Grid } from '@material-ui/core'
-import { Skeleton } from '@material-ui/lab'
-import { withStyles } from '@material-ui/core/styles'
+import { Box, Typography, Grid, Skeleton } from "@mui/material";
 
-import { getHome } from '../../../utils/getters'
-import Header from '../../../components/Header'
+import { getHome } from "../../../utils/getters";
+import Header from "../../../components/Header";
 
-export default withStyles({
-  root: {
-    color: 'white'
-  }
-})(({ data, classes }) => {
+export default function Main({ data }) {
   return (
     <Box p={{ xs: 0, md: 1 }}>
       <Head>
@@ -26,16 +19,16 @@ export default withStyles({
         <Typography>Private messages</Typography>
       </Box>
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getHome(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getHome(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
