@@ -1,24 +1,36 @@
-import React from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
+import React from "react";
+import Head from "next/head";
+import Link from "next/link";
 
-import { Box, Grid, Typography, Hidden, Accordion, AccordionSummary, AccordionDetails, List, ListItem, ListItemText } from '@material-ui/core'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
-import { withStyles } from '@material-ui/core/styles'
+import {
+  Box,
+  Grid,
+  Typography,
+  Hidden,
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
+  List,
+  ListItem,
+  ListItemText,
+} from "@mui/material";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import { styled } from "@mui/system";
+import { useTheme } from "@mui/material/styles";
 
-import { getForum } from '../../../utils/getters'
-import Header from '../../../components/Header'
-import Pagination from '../../../components/Pagination'
-import { formatDate } from '../../../utils/formatters'
+import { getForum } from "../../../utils/getters";
+import Header from "../../../components/Header";
+import Pagination from "../../../components/Pagination";
+import { formatDate } from "../../../utils/formatters";
 
-import theme from '../../../utils/theme'
+const StyledTypography = styled(Typography)({
+  color: "white",
+});
 
-export default withStyles({
-  root: {
-    color: 'white'
-  }
-})(({ data, classes }) => {
-  console.log(data)
+export default function Main({ data }) {
+  const theme = useTheme();
+
+  console.log(data);
 
   return (
     <Box p={{ xs: 0, md: 1 }}>
@@ -34,37 +46,27 @@ export default withStyles({
             <Grid container>
               <Grid item xs={7}>
                 <Box p={1}>
-                  <Typography align="center" className={classes.root}>
-                    Topics
-                  </Typography>
+                  <StyledTypography align="center">Topics</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={1}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Replies
-                  </Typography>
+                  <StyledTypography align="right">Replies</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={1}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Author
-                  </Typography>
+                  <StyledTypography align="right">Author</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={1}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Views
-                  </Typography>
+                  <StyledTypography align="right">Views</StyledTypography>
                 </Box>
               </Grid>
               <Grid item xs={2}>
                 <Box p={1}>
-                  <Typography align="right" className={classes.root}>
-                    Last Post
-                  </Typography>
+                  <StyledTypography align="right">Last Post</StyledTypography>
                 </Box>
               </Grid>
             </Grid>
@@ -72,12 +74,19 @@ export default withStyles({
         </Hidden>
 
         {data.page.forumGroups.map((el, i) => {
-          if (el.title.toLowerCase().includes('sticky')) {
+          if (el.title.toLowerCase().includes("sticky")) {
             return (
               <Box mb={2} key={`forumGroup-${i}`}>
-                <Accordion square={true} key={`classifieds-categories-${i}-mobile`}>
-                  <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls={`panel${i}bh-content`} id={`panel${i}bh-header`}>
-                    <Typography className={classes.heading}>
+                <Accordion
+                  square={true}
+                  key={`classifieds-categories-${i}-mobile`}
+                >
+                  <AccordionSummary
+                    expandIcon={<ExpandMoreIcon />}
+                    aria-controls={`panel${i}bh-content`}
+                    id={`panel${i}bh-header`}
+                  >
+                    <Typography>
                       {el.title} ({el.items.length})
                     </Typography>
                   </AccordionSummary>
@@ -85,15 +94,29 @@ export default withStyles({
                     <Box width="100%">
                       <List disablePadding={true}>
                         {el.items.map((sub, subIdx) => (
-                          <ListItem button style={{ alignItems: 'flex-start', border: `1px solid ${theme.palette.secondary.light}`, borderWidth: '0 0 1px' }} key={`classifieds-subcategories-${subIdx}-mobile`}>
+                          <ListItem
+                            button
+                            style={{
+                              alignItems: "flex-start",
+                              border: `1px solid ${theme.palette.secondary.light}`,
+                              borderWidth: "0 0 1px",
+                            }}
+                            key={`classifieds-subcategories-${subIdx}-mobile`}
+                          >
                             <ListItemText>
                               <Link href={sub.href} passHref>
-                                <Typography component="a" className={classes.heading}>
-                                  {sub.title}
-                                </Typography>
+                                {sub.title}
                               </Link>
                             </ListItemText>
-                            <ListItemText style={{ textAlign: 'right', whiteSpace: 'nowrap', flex: '0 0 80px' }}>{formatDate(sub.lastPost.text).short}</ListItemText>
+                            <ListItemText
+                              style={{
+                                textAlign: "right",
+                                whiteSpace: "nowrap",
+                                flex: "0 0 80px",
+                              }}
+                            >
+                              {formatDate(sub.lastPost.text).short}
+                            </ListItemText>
                           </ListItem>
                         ))}
                       </List>
@@ -101,7 +124,7 @@ export default withStyles({
                   </AccordionDetails>
                 </Accordion>
               </Box>
-            )
+            );
           }
 
           return (
@@ -117,10 +140,12 @@ export default withStyles({
                         <Box mb={{ xs: 0, sm: 0 }}>
                           {/* {subEl.newPosts ? '!' : ''} */}
 
-                          <Link href={subEl.href || ''} passHref>
-                            <Typography component="a" className="font-semibold">
-                              {subEl.title}
-                            </Typography>
+                          <Link
+                            href={subEl.href || ""}
+                            passHref
+                            className="font-semibold"
+                          >
+                            {subEl.title}
                           </Link>
                           <Typography>{subEl.description}</Typography>
                         </Box>
@@ -130,18 +155,20 @@ export default withStyles({
                           <Typography align="right">{subEl.replies}</Typography>
                         </Grid>
                         <Grid item xs={5} sm={1} align="right">
-                          <Typography align="right">{subEl.author.text}</Typography>
-                          <Link href={subEl.author.user.href}>
-                            <Typography component="a" align="right">
-                              {subEl.author.user.title}
-                            </Typography>
+                          <Typography align="right">
+                            {subEl.author.text}
+                          </Typography>
+                          <Link href={subEl.author.user.href} align="right">
+                            {subEl.author.user.title}
                           </Link>
                         </Grid>
                         <Grid item xs={2} sm={1}>
                           <Typography align="right">{subEl.views}</Typography>
                         </Grid>
                         <Grid item xs={3} sm={2}>
-                          <Typography align="right">{subEl.lastPost.text}</Typography>
+                          <Typography align="right">
+                            {subEl.lastPost.text}
+                          </Typography>
                         </Grid>
                       </Hidden>
                     </Grid>
@@ -149,22 +176,22 @@ export default withStyles({
                 ))}
               </Box>
             </Box>
-          )
+          );
         })}
       </Box>
 
       <Pagination pages={data.page.pages} />
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getForum(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getForum(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
