@@ -1,57 +1,37 @@
-import React, { useState } from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
-import Image from 'next/image'
+import React, { useState } from "react";
+import Head from "next/head";
+import Link from "next/link";
 
-import { Box, Typography, Grid, Accordion, AccordionSummary, AccordionDetails, List, ListItem, ListItemText, Hidden, IconButton, Paper } from '@material-ui/core'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
-import SearchIcon from '@material-ui/icons/Search'
-import { withStyles } from '@material-ui/core/styles'
+import {
+  Box,
+  Typography,
+  Grid,
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
+  List,
+  ListItem,
+  ListItemText,
+  Hidden,
+  IconButton,
+  Paper,
+} from "@mui/material";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import SearchIcon from "@mui/icons-material/Search";
+import { useTheme } from "@mui/material/styles";
 
-import { getClassifieds } from '../../../utils/getters'
-import Header from '../../../components/Header'
-import ClassifiedsSearch from '../../../components/ClassifiedsSearch'
+import { getClassifieds } from "../../../utils/getters";
+import Header from "../../../components/Header";
+import ClassifiedsSearch from "../../../components/ClassifiedsSearch";
 
-export default withStyles((theme) => ({
-  root: {
-    color: 'white'
-  },
-  list: {
-    width: 250
-  },
+export default function Main({ data }) {
+  const [showSearch, setShowSearch] = useState(false);
 
-  accRoot: {
-    width: '100%'
-  },
-  heading: {
-    display: 'block',
-    fontSize: theme.typography.pxToRem(15),
-    flexShrink: 0
-  },
-
-  popper: {
-    minWidth: 200
-  },
-
-  menuItem: {
-    justifyContent: 'center'
-  },
-
-  menuBtn: {
-    height: 64
-  },
-
-  searchBtn: {
-    width: '40px',
-    padding: '14px 0',
-    textAlign: 'center'
-  }
-}))(({ data, classes }) => {
-  const [showSearch, setShowSearch] = useState(false)
+  const theme = useTheme();
 
   const handleShowSearch = () => {
-    setShowSearch(!showSearch)
-  }
+    setShowSearch(!showSearch);
+  };
 
   return (
     <Box p={{ xs: 0, md: 1 }}>
@@ -65,7 +45,12 @@ export default withStyles((theme) => ({
         <Grid container justifyContent="space-between">
           <Hidden smDown>
             <Grid item xs={11} md={3}>
-              <Box mb={1} mr={{ md: 2 }} border={1} borderColor="secondary.light">
+              <Box
+                mb={1}
+                mr={{ md: 2 }}
+                border={1}
+                borderColor="secondary.light"
+              >
                 <Box bgcolor="secondary.light" p={1}>
                   <Typography>Categories</Typography>
                 </Box>
@@ -77,9 +62,11 @@ export default withStyles((theme) => ({
                       </Box>
                       <Box mb={3}>
                         {el.items.map((subEl, subIdx) => (
-                          <Box key={`classifieds-subcategories-${subIdx}-desktop`}>
+                          <Box
+                            key={`classifieds-subcategories-${subIdx}-desktop`}
+                          >
                             <Link href={subEl.href} passHref>
-                              <Typography component="a">{subEl.title}</Typography>
+                              {subEl.title}
                             </Link>
                           </Box>
                         ))}
@@ -90,7 +77,7 @@ export default withStyles((theme) => ({
               </Box>
             </Grid>
             <Grid item xs={1}>
-              <IconButton aria-label="search" className={classes.margin} size="small">
+              <IconButton aria-label="search" size="small">
                 <SearchIcon fontSize="inherit" />
               </IconButton>
             </Grid>
@@ -101,25 +88,58 @@ export default withStyles((theme) => ({
               <Box mb={2} pr={6} position="relative">
                 <Accordion square={true}>
                   <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                    <Typography className={classes.heading}>Categories</Typography>
+                    <Typography
+                      sx={{
+                        display: "block",
+                        fontSize: theme.typography.pxToRem(15),
+                        flexShrink: 0,
+                      }}
+                    >
+                      Categories
+                    </Typography>
                   </AccordionSummary>
                   <AccordionDetails direction="column">
                     <Box width="100%">
                       {data.page.categories.map((el, i) => (
-                        <Accordion square={true} key={`classifieds-categories-${i}-mobile`}>
-                          <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls={`panel${i}bh-content`} id={`panel${i}bh-header`}>
-                            <Typography className={classes.heading}>{el.title}</Typography>
+                        <Accordion
+                          square={true}
+                          key={`classifieds-categories-${i}-mobile`}
+                        >
+                          <AccordionSummary
+                            expandIcon={<ExpandMoreIcon />}
+                            aria-controls={`panel${i}bh-content`}
+                            id={`panel${i}bh-header`}
+                          >
+                            <Typography
+                              sx={{
+                                display: "block",
+                                fontSize: theme.typography.pxToRem(15),
+                                flexShrink: 0,
+                              }}
+                            >
+                              {el.title}
+                            </Typography>
                           </AccordionSummary>
                           <AccordionDetails>
                             <Box width="100%">
                               <List disablePadding={true}>
                                 {el.items.map((sub, subIdx) => (
-                                  <ListItem button key={`classifieds-subcategories-${subIdx}-mobile`}>
+                                  <ListItem
+                                    button
+                                    key={`classifieds-subcategories-${subIdx}-mobile`}
+                                  >
                                     <ListItemText>
-                                      <Link href={sub.href} passHref>
-                                        <Typography component="a" className={classes.heading}>
-                                          {sub.title}
-                                        </Typography>
+                                      <Link
+                                        href={sub.href}
+                                        passHref
+                                        sx={{
+                                          display: "block",
+                                          fontSize:
+                                            theme.typography.pxToRem(15),
+                                          flexShrink: 0,
+                                        }}
+                                      >
+                                        {sub.title}
                                       </Link>
                                     </ListItemText>
                                   </ListItem>
@@ -134,7 +154,16 @@ export default withStyles((theme) => ({
                 </Accordion>
                 <Box position="absolute" right="0" top="0">
                   <Paper variant="outlined" onClick={handleShowSearch}>
-                    <IconButton aria-label="search" className={classes.searchBtn} size="small" variant="outlined">
+                    <IconButton
+                      aria-label="search"
+                      sx={{
+                        width: "40px",
+                        padding: "14px 0",
+                        textAlign: "center",
+                      }}
+                      size="small"
+                      variant="outlined"
+                    >
                       <SearchIcon fontSize="default" />
                     </IconButton>
                   </Paper>
@@ -152,15 +181,22 @@ export default withStyles((theme) => ({
               <Box p={1}>
                 <Grid container justifyContent="center" alignItems="flex-start">
                   {data.page.featuredAds.map((el, i) => (
-                    <Grid item xs={6} sm={4} md={2} key={`classifieds-random-ads-${i}`}>
+                    <Grid
+                      item
+                      xs={6}
+                      sm={4}
+                      md={2}
+                      key={`classifieds-random-ads-${i}`}
+                    >
                       <Box px={2} pb={2} align="center">
-                        <Link href={el.href} key={`classifieds-featured-ads-${i}`}>
-                          <a>
-                            <div>
-                              <img src={el.img.src} alt={el.title} />
-                            </div>
-                            <Typography align="center">{el.title}</Typography>
-                          </a>
+                        <Link
+                          href={el.href}
+                          key={`classifieds-featured-ads-${i}`}
+                        >
+                          <div>
+                            <img src={el.img.src} alt={el.title} />
+                          </div>
+                          <Typography align="center">{el.title}</Typography>
                         </Link>
                       </Box>
                     </Grid>
@@ -177,17 +213,27 @@ export default withStyles((theme) => ({
               <Box p={1}>
                 <Grid container justifyContent="center" alignItems="flex-start">
                   {data.page.randomAds.map((el, i) => (
-                    <Grid item xs={6} sm={4} md={2} key={`classifieds-random-ads-${i}`}>
+                    <Grid
+                      item
+                      xs={6}
+                      sm={4}
+                      md={2}
+                      key={`classifieds-random-ads-${i}`}
+                    >
                       <Box px={2} pb={2} align="center">
                         <Link href={el.href}>
-                          <a>
-                            <div>
-                              <img src={el.img.src} alt={el.title} />
-                            </div>
-                            <Typography component="div" align="center">
-                              <Box {...(el.heavyTitle && { fontWeight: 'fontWeightBold' })}>{el.title}</Box>
-                            </Typography>
-                          </a>
+                          <div>
+                            <img src={el.img.src} alt={el.title} />
+                          </div>
+                          <Typography component="div" align="center">
+                            <Box
+                              {...(el.heavyTitle && {
+                                fontWeight: "fontWeightBold",
+                              })}
+                            >
+                              {el.title}
+                            </Box>
+                          </Typography>
                         </Link>
                       </Box>
                     </Grid>
@@ -201,16 +247,16 @@ export default withStyles((theme) => ({
 
       <ClassifiedsSearch open={showSearch} setOpen={setShowSearch} />
     </Box>
-  )
-})
+  );
+}
 
 export async function getServerSideProps(context) {
-  const { data, ...rest } = await getClassifieds(context.req)
-  context.res.setHeader('set-cookie', rest.cookies || [])
+  const { data, ...rest } = await getClassifieds(context.req);
+  context.res.setHeader("set-cookie", rest.cookies || []);
 
   return {
     props: {
-      data
-    }
-  }
+      data,
+    },
+  };
 }
