@@ -1,23 +1,23 @@
-import { useEffect, useState } from 'react'
-import { useRouter } from 'next/router'
-import ReactGA from 'react-ga'
+import { useEffect, useState } from "react";
+import { useRouter } from "next/router";
+import ReactGA from "react-ga";
 
 const usePageTracking = () => {
-  const { asPath } = useRouter()
-  const [initialized, setInitialized] = useState(false)
+  const { asPath } = useRouter();
+  const [initialized, setInitialized] = useState(false);
 
   useEffect(() => {
-    if (!window.location.href.includes('localhost')) {
-      ReactGA.initialize('UA-194225978-1')
-      setInitialized(true)
+    if (!window.location.href.includes("localhost")) {
+      ReactGA.initialize("UA-194225978-1");
+      setInitialized(true);
     }
-  }, [])
+  }, []);
 
   useEffect(() => {
     if (initialized === true) {
-      ReactGA.pageview(asPath)
+      ReactGA.pageview(asPath);
     }
-  }, [initialized, asPath])
-}
+  }, [initialized, asPath]);
+};
 
-export default usePageTracking
+export default usePageTracking;
